Fix src

+++ spark/src/main/scala/org/apache/spark/sql/comet/CometBatchScanExec.scala
@@ -70,7 +70,7 @@ case class CometBatchScanExec(wrapped: BatchScanExec, runtimeFilters: Seq[Expres
 
         override def next(): ColumnarBatch = {
           val batch = batches.next()
-          numOutputRows += batch.numRows()
+          numOutputRows += batch.numRows().toLong
           batch
         }
       }
